fix(core): store if we trigger the torrent finished signal (#247)
The old check that just checks `total_download` is not robust enough and
can lead to multiple invocations in some cases. Instead, we use the
torrent metadata to store if we triggered the torrent finished signal
and check that. This should lead to the signal only being triggered once
for the lifetime of the torrent.